Finished all part1 exercises
Finished all the part1 exercises now, but with Anecdotes I was a bit lost on the cleanest way to write the code out in terms of best practice. Was not sure if I should create a new component for the text, and how to best add new lines aside from using divs.